ДЗ6. Добавлен AnimalsRepository, подключен Spring, AnimalsRepository и CreateAnimalService - бины. Добавлен постпроцессор.
