+++ build.gradle
@@ -12,6 +12,10 @@ repositories {
 dependencies {
     testImplementation platform('org.junit:junit-bom:5.10.0')
     testImplementation 'org.junit.jupiter:junit-jupiter:5.10.0'
+    implementation 'org.springframework:spring-core:6.1.3'
+    implementation 'org.springframework:spring-context:6.1.3'
+    implementation 'org.springframework:spring-beans:6.1.3'
+    implementation 'javax.annotation:javax.annotation-api:1.3.2'
 }
 
 test {
