+++ src/main/java/ru/mts/Main.java
@@ -1,44 +1,39 @@
 package ru.mts;
 
-import ru.mts.hw_4.animal.Animal;
-import ru.mts.hw_4.animal.impl.Rabbit;
-import ru.mts.hw_4.animal.impl.Wolf;
-import ru.mts.hw_4.service.SearchService;
-import ru.mts.hw_4.service.impl.CreateAnimalServiceImpl;
-import ru.mts.hw_4.service.impl.SearchServiceImpl;
-
-import java.math.BigDecimal;
-import java.time.LocalDate;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+import ru.mts.hw_6.animal.Animal;
+import ru.mts.hw_6.config.Config;
+import ru.mts.hw_6.repository.AnimalsRepository;
+import ru.mts.hw_6.service.CreateAnimalService;
+import ru.mts.hw_6.service.impl.CreateAnimalServiceImpl;
 
 
 public class Main {
     public static void main(String[] args) {
-        CreateAnimalServiceImpl createService = new CreateAnimalServiceImpl();
-        SearchService searchService = new SearchServiceImpl();
-        Animal[] arrayOfDuplicatedAnimals = {new Wolf("gray", BigDecimal.valueOf(15), LocalDate.now()),
-                new Wolf("gray", BigDecimal.valueOf(15), LocalDate.now()),
-                new Wolf("gray", BigDecimal.valueOf(15), LocalDate.now()),
-                new Rabbit("white", BigDecimal.valueOf(15), LocalDate.now()),
-                new Rabbit("white", BigDecimal.valueOf(15), LocalDate.now())
-        };
+        ApplicationContext context = new AnnotationConfigApplicationContext(Config.class);
+        AnimalsRepository animalsRepository = context.getBean(AnimalsRepository.class);
 
-        Animal[] animalArray = createService.createAnimals(5);
-        String[] animalNamesArray = searchService.findLeapYearNames(animalArray);
+        String[] animalNamesArray = animalsRepository.findLeapYearNames();
         System.out.println("Животные, родившиеся в високосный год:");
         for (String animalName : animalNamesArray) {
             System.out.println(animalName);
         }
         System.out.println("-----");
-        Animal[] olderAnimalsArray = searchService.findOlderAnimals(animalArray, 1);
+        Animal[] olderAnimalsArray = animalsRepository.findOlderAnimals(1);
         System.out.println("Животные старше 1 года:");
         for (Animal animal : olderAnimalsArray) {
             System.out.println(animal);
         }
         System.out.println("-----");
-        System.out.println("Поиск одинаковых животные в массиве с разными животными:");
-        searchService.printDuplicates(animalArray);
+        System.out.println("Поиск одинаковых животных:");
+        animalsRepository.printDuplicates();
 
-        System.out.println("Поиск одинаковых животных в массиве с дубликатами:");
-        searchService.printDuplicates(arrayOfDuplicatedAnimals);
+        System.out.println("Bean прототип");
+        CreateAnimalService createService = context.getBean(CreateAnimalServiceImpl.class);
+        CreateAnimalService anotherCreateService = context.getBean(CreateAnimalServiceImpl.class);
+        System.out.println(createService.getAnimal());
+        System.out.println(anotherCreateService.getAnimal());
+        System.out.println("-----");
     }
 }
