--- src/main/java/ru/mts/hw_4/factory/AnimalTypes.java
@@ -1,8 +1,0 @@
-package ru.mts.hw_4.factory;
-
-public enum AnimalTypes {
-    WOLF,
-    SHARK,
-    RABBIT,
-    DEER
-}
